+++ qirlib/src/utils.rs
